fix: bucket вынести конструкторы в отдельный файл new.go
